Factor out PosOccFieldParams and PosOccFieldsParams from posocccompression compilation unit.
