+++ searchlib/src/vespa/searchlib/bitcompression/posocccompression.cpp
@@ -2,6 +2,7 @@
 
 #include "compression.h"
 #include "posocccompression.h"
+#include "posocc_fields_params.h"
 #include <vespa/searchlib/index/schemautil.h>
 #include <vespa/searchlib/fef/termfieldmatchdata.h>
 #include <vespa/searchlib/fef/termfieldmatchdataarray.h>
@@ -29,295 +30,6 @@ vespalib::string EG64PosOccId2 = "EG64PosOcc.2";    // Fixed k values
 }
 
 namespace search::bitcompression {
-
-
-PosOccFieldParams::PosOccFieldParams()
-    : _elemLenK(0),
-      _hasElements(false),
-      _hasElementWeights(false),
-      _avgElemLen(512),
-      _collectionType(SINGLE),
-      _name()
-{ }
-
-
-bool
-PosOccFieldParams::operator==(const PosOccFieldParams &rhs) const
-{
-    return _collectionType == rhs._collectionType &&
-              _avgElemLen == rhs._avgElemLen &&
-                     _name == rhs._name;
-}
-
-
-vespalib::string
-PosOccFieldParams::getParamsPrefix(uint32_t idx)
-{
-    vespalib::asciistream paramsPrefix;
-    paramsPrefix << "fieldParams.";
-    paramsPrefix << idx;
-    return paramsPrefix.str();
-}
-
-
-void
-PosOccFieldParams::getParams(PostingListParams &params, uint32_t idx) const
-{
-    vespalib::string paramsPrefix = getParamsPrefix(idx);
-    vespalib::string collStr = paramsPrefix + ".collectionType";
-    vespalib::string avgElemLenStr = paramsPrefix + ".avgElemLen";
-    vespalib::string nameStr = paramsPrefix + ".name";
-
-    switch (_collectionType) {
-    case SINGLE:
-        params.setStr(collStr, "single");
-        break;
-    case ARRAY:
-        params.setStr(collStr, "array");
-        break;
-    case WEIGHTEDSET:
-        params.setStr(collStr, "weightedSet");
-        break;
-    }
-    params.set(avgElemLenStr, _avgElemLen);
-    params.setStr(nameStr, _name);
-}
-
-
-void
-PosOccFieldParams::setParams(const PostingListParams &params, uint32_t idx)
-{
-    vespalib::string paramsPrefix = getParamsPrefix(idx);
-    vespalib::string collStr = paramsPrefix + ".collectionType";
-    vespalib::string avgElemLenStr = paramsPrefix + ".avgElemLen";
-    vespalib::string nameStr = paramsPrefix + ".name";
-
-    if (params.isSet(collStr)) {
-        vespalib::string collVal = params.getStr(collStr);
-        if (collVal == "single") {
-            _collectionType = SINGLE;
-            _hasElements = false;
-            _hasElementWeights = false;
-        } else if (collVal == "array") {
-            _collectionType = ARRAY;
-            _hasElements = true;
-            _hasElementWeights = false;
-        } else if (collVal == "weightedSet") {
-            _collectionType = WEIGHTEDSET;
-            _hasElements = true;
-            _hasElementWeights = true;
-        }
-    }
-    params.get(avgElemLenStr, _avgElemLen);
-    if (params.isSet(nameStr)) {
-        _name = params.getStr(nameStr);
-    }
-}
-
-
-void
-PosOccFieldParams::setSchemaParams(const Schema &schema, uint32_t fieldId)
-{
-    assert(fieldId < schema.getNumIndexFields());
-    const Schema::IndexField &field = schema.getIndexField(fieldId);
-    switch (field.getCollectionType()) {
-    case schema::CollectionType::SINGLE:
-        _collectionType = SINGLE;
-        _hasElements = false;
-        _hasElementWeights = false;
-        break;
-    case schema::CollectionType::ARRAY:
-        _collectionType = ARRAY;
-        _hasElements = true;
-        _hasElementWeights = false;
-        break;
-    case schema::CollectionType::WEIGHTEDSET:
-        _collectionType = WEIGHTEDSET;
-        _hasElements = true;
-        _hasElementWeights = true;
-        break;
-    default:
-        LOG(error, "Bad collection type");
-        LOG_ABORT("should not be reached");
-    }
-    _avgElemLen = field.getAvgElemLen();
-    _name = field.getName();
-}
-
-
-void
-PosOccFieldParams::readHeader(const vespalib::GenericHeader &header,
-                              const vespalib::string &prefix)
-{
-    vespalib::string nameKey(prefix + "fieldName");
-    vespalib::string collKey(prefix + "collectionType");
-    vespalib::string avgElemLenKey(prefix + "avgElemLen");
-    _name = header.getTag(nameKey).asString();
-    Schema::CollectionType ct = schema::collectionTypeFromName(header.getTag(collKey).asString());
-    switch (ct) {
-    case schema::CollectionType::SINGLE:
-        _collectionType = SINGLE;
-        _hasElements = false;
-        _hasElementWeights = false;
-        break;
-    case schema::CollectionType::ARRAY:
-        _collectionType = ARRAY;
-        _hasElements = true;
-        _hasElementWeights = false;
-        break;
-    case schema::CollectionType::WEIGHTEDSET:
-        _collectionType = WEIGHTEDSET;
-        _hasElements = true;
-        _hasElementWeights = true;
-        break;
-    default:
-        LOG_ABORT("Bad collection type when reading field param in header");
-    }
-    _avgElemLen = header.getTag(avgElemLenKey).asInteger();
-}
-
-
-void
-PosOccFieldParams::writeHeader(vespalib::GenericHeader &header,
-                               const vespalib::string &prefix) const
-{
-    vespalib::string nameKey(prefix + "fieldName");
-    vespalib::string collKey(prefix + "collectionType");
-    vespalib::string avgElemLenKey(prefix + "avgElemLen");
-    header.putTag(GenericHeader::Tag(nameKey, _name));
-    Schema::CollectionType ct(schema::CollectionType::SINGLE);
-    switch (_collectionType) {
-    case SINGLE:
-        ct = schema::CollectionType::SINGLE;
-        break;
-    case ARRAY:
-        ct = schema::CollectionType::ARRAY;
-        break;
-    case WEIGHTEDSET:
-        ct = schema::CollectionType::WEIGHTEDSET;
-        break;
-    default:
-        LOG_ABORT("Bad collection type when writing field param in header");
-    }
-    header.putTag(GenericHeader::Tag(collKey, schema::getTypeName(ct)));
-    header.putTag(GenericHeader::Tag(avgElemLenKey, _avgElemLen));
-}
-
-
-PosOccFieldsParams::PosOccFieldsParams()
-    : _numFields(0u),
-      _fieldParams(nullptr),
-      _params()
-{
-}
-
-PosOccFieldsParams::PosOccFieldsParams(const PosOccFieldsParams &rhs)
-    : _numFields(0u),
-      _fieldParams(nullptr),
-      _params(rhs._params)
-{
-    cacheParamsRef();
-}
-
-PosOccFieldsParams &
-PosOccFieldsParams::operator=(const PosOccFieldsParams &rhs)
-{
-    assertCachedParamsRef();
-    _params = rhs._params;
-    cacheParamsRef();
-    return *this;
-}
-
-
-bool
-PosOccFieldsParams::operator==(const PosOccFieldsParams &rhs) const
-{
-    return _params == rhs._params;
-}
-
-
-void
-PosOccFieldsParams::getParams(PostingListParams &params) const
-{
-    assertCachedParamsRef();
-    assert(_numFields == 1u); // Only single field for now
-    params.set("numFields", _numFields);
-    // Single posting file index format will have multiple fields in file
-    for (uint32_t field = 0; field < _numFields; ++field) {
-        _fieldParams[field].getParams(params, field);
-    }
-}
-
-
-void
-PosOccFieldsParams::setParams(const PostingListParams &params)
-{
-    assertCachedParamsRef();
-    uint32_t numFields = _numFields;
-    params.get("numFields", numFields);
-    assert(numFields == 1u);
-    _params.resize(numFields);
-    cacheParamsRef();
-    // Single posting file index format will have multiple fields in file
-    for (uint32_t field = 0; field < numFields; ++field) {
-        _params[field].setParams(params, field);
-    }
-}
-
-
-void
-PosOccFieldsParams::setSchemaParams(const Schema &schema,
-                                    const uint32_t indexId)
-{
-    assertCachedParamsRef();
-    SchemaUtil::IndexIterator i(schema, indexId);
-    assert(i.isValid());
-    _params.resize(1u);
-    cacheParamsRef();
-    const Schema::IndexField &field = schema.getIndexField(indexId);
-    if (!SchemaUtil::validateIndexField(field)) {
-        LOG_ABORT("should not be reached");
-    }
-    _params[0].setSchemaParams(schema, indexId);
-}
-
-
-void
-PosOccFieldsParams::readHeader(const vespalib::GenericHeader &header,
-               const vespalib::string &prefix)
-{
-    vespalib::string numFieldsKey(prefix + "numFields");
-    assertCachedParamsRef();
-    uint32_t numFields = header.getTag(numFieldsKey).asInteger();
-    assert(numFields == 1u);
-    _params.resize(numFields);
-    cacheParamsRef();
-    // Single posting file index format will have multiple fields in file
-    for (uint32_t field = 0; field < numFields; ++field) {
-        vespalib::asciistream as;
-        as << prefix << "field[" << field << "].";
-        vespalib::string subPrefix(as.str());
-        _params[field].readHeader(header, subPrefix);
-    }
-}
-
-
-void
-PosOccFieldsParams::writeHeader(vespalib::GenericHeader &header,
-                                const vespalib::string &prefix) const
-{
-    vespalib::string numFieldsKey(prefix + "numFields");
-    assertCachedParamsRef();
-    assert(_numFields == 1u);
-    header.putTag(GenericHeader::Tag(numFieldsKey, _numFields));
-    // Single posting file index format will have multiple fields in file
-    for (uint32_t field = 0; field < _numFields; ++field) {
-        vespalib::asciistream as;
-        as << prefix << "field[" << field << "].";
-        vespalib::string subPrefix(as.str());
-        _params[field].writeHeader(header, subPrefix);
-    }
-}
 
 
 template <bool bigEndian>
